2014/11/2

+++ DragEditor/MySourceData.m
@@ -32,7 +32,7 @@
 
 -(int)count
 {
-    return [data count];
+    return (int)([data count]);
 }
 
 -(id)init
